dashboard/requirements.txt
streamlit
pandas
requests
plotly
scipy
python-dotenv
streamlit-navigation-bar
matplotlib
setuptools
wheel